feat: refactored code and updated deps
