+++ .gitignore
@@ -1,6 +1,4 @@
 node_modules
 dist
-package-lock.json
-yarn-error.log
 .reports
 .docs
