--- .huskyrc
@@ -1,8 +1,0 @@
-{
-    "hooks": {
-        "pre-push": "yarn test",
-        "pre-commit": "yarn lint",
-        "commit-msg": "commitlint -E HUSKY_GIT_PARAMS",
-        "prepare-commit-msg": "exec < /dev/tty && git cz --hook || true"
-    }
-}
