+++ CHANGELOG.md
@@ -2,20 +2,20 @@
 
 ### Bug Fixes
 
--   fixing the release action ([92e5584](https://github.com/nass600/plex-sdk/commit/92e5584e670d1a44aff62f1e3caefa5af9a3c29f))
+*   fixing the release action ([92e5584](https://github.com/nass600/plex-sdk/commit/92e5584e670d1a44aff62f1e3caefa5af9a3c29f))
 
 # [2.0.0](https://github.com/nass600/plex-sdk/compare/1.0.5...2.0.0) (2023-01-09)
 
 ### Features
 
--   modified search endpoint to accept optional params ([400e347](https://github.com/nass600/plex-sdk/commit/400e347008c45fc3df77ff624d4f1d9972d4a477))
+*   modified search endpoint to accept optional params ([400e347](https://github.com/nass600/plex-sdk/commit/400e347008c45fc3df77ff624d4f1d9972d4a477))
 
 ### BREAKING CHANGES
 
--   The pms.search function modifies the signature to accept and object, including the
+*   The pms.search function modifies the signature to accept and object, including the
     query
 
--   ci: changing to github actions
+*   ci: changing to github actions
 
 ## [1.0.5](https://github.com/nass600/plex-sdk/compare/1.0.4...1.0.5) (2021-01-10)
 
@@ -25,26 +25,26 @@
 
 ### Bug Fixes
 
--   cleaned up documentation files ([1c9650c](https://github.com/nass600/plex-sdk/commit/1c9650cabf2b46e934b73b5758fa4c9915c7eda1))
+*   cleaned up documentation files ([1c9650c](https://github.com/nass600/plex-sdk/commit/1c9650cabf2b46e934b73b5758fa4c9915c7eda1))
 
 ## [1.0.2](https://github.com/nass600/plex-sdk/compare/1.0.1...1.0.2) (2021-01-02)
 
 ### Bug Fixes
 
--   fixed badges in readme ([c002f6b](https://github.com/nass600/plex-sdk/commit/c002f6b0455762ac5a0ecdffe845a259697ea545))
+*   fixed badges in readme ([c002f6b](https://github.com/nass600/plex-sdk/commit/c002f6b0455762ac5a0ecdffe845a259697ea545))
 
 ## [1.0.1](https://github.com/nass600/plex-sdk/compare/1.0.0...1.0.1) (2021-01-02)
 
 ### Bug Fixes
 
--   removed moment ([a47aa14](https://github.com/nass600/plex-sdk/commit/a47aa142a946dc812b2ec30d118f16651854c3f2))
+*   removed moment ([a47aa14](https://github.com/nass600/plex-sdk/commit/a47aa142a946dc812b2ec30d118f16651854c3f2))
 
 # 1.0.0 (2021-01-02)
 
 ### Bug Fixes
 
--   fixed semantic release config ([d2d46fd](https://github.com/nass600/plex-sdk/commit/d2d46fd25cbd55f7820c51bd3485a6ebd3524922))
+*   fixed semantic release config ([d2d46fd](https://github.com/nass600/plex-sdk/commit/d2d46fd25cbd55f7820c51bd3485a6ebd3524922))
 
 ### Features
 
--   initial endpoints ([40e4d05](https://github.com/nass600/plex-sdk/commit/40e4d0520bde0aee6309379f958e99ee5f28a105))
+*   initial endpoints ([40e4d05](https://github.com/nass600/plex-sdk/commit/40e4d0520bde0aee6309379f958e99ee5f28a105))
