+++ commitlint.config.ts
@@ -1,3 +1,3 @@
-module.exports = {
+export default {
     extends: ['@commitlint/config-conventional']
 }
