+++ package.json
@@ -3,31 +3,32 @@
     "version": "2.0.1",
     "description": "Typescript SDK for Plex and Plex Media Server APIs",
     "repository": "git@github.com:nass600/plex-sdk",
-    "author": "Ignacio Velazquez <ivelazquez85@gmail.com>",
     "license": "MIT",
+    "author": "Ignacio Velazquez <ivelazquez85@gmail.com>",
     "main": "dist/plex.js",
     "module": "dist/plex.esm.js",
     "types": "dist/index.d.ts",
     "scripts": {
-        "type-check": "tsc --noEmit",
-        "type-check:watch": "yarn type-check --watch",
-        "prelint": "yarn type-check",
-        "lint": "yarn run lint:js && yarn run lint:md",
-        "lint:js": "eslint . -c .eslintrc.js --ext .js,.jsx,.ts,.tsx --ignore-path .eslintignore --fix",
-        "lint:md": "remark ./*.md -o",
-        "pretest": "yarn type-check",
-        "test": "jest --config jest.config.js",
-        "prebuild": "yarn type-check",
-        "build": "rollup -c",
-        "build:watch": "rollup -c -w",
+        "prebuild": "npm run type-check",
+        "build": "rollup -c --configPlugin typescript",
         "docs": "typedoc",
-        "prepublishOnly": "yarn build && yarn docs",
-        "postversion": "yarn lint:md"
+        "prelint": "npm run type-check",
+        "lint": "npm run lint:js && npm run lint:md",
+        "lint:js": "eslint . --ext ts,js --fix",
+        "lint:md": "remark ./*.md -o",
+        "prepare": "husky install",
+        "prepublishOnly": "npm run build && npm run docs",
+        "test": "jest --config jest.config.ts",
+        "type-check": "tsc --noEmit",
+        "postversion": "npm run lint:md"
+    },
+    "config": {
+        "commitizen": {
+            "path": "cz-conventional-changelog"
+        }
     },
     "dependencies": {
-        "axios": "0.19.2",
-        "bluebird": "^3.7.2",
-        "js-base64": "2.5.1"
+        "axios": "^1.2.2"
     },
     "devDependencies": {
         "@babel/core": "7.7.7",
@@ -39,66 +40,39 @@
         "@babel/preset-typescript": "7.7.7",
         "@babel/register": "7.7.7",
         "@babel/runtime": "7.7.7",
-        "@commitlint/cli": "^11.0.0",
-        "@commitlint/config-conventional": "^11.0.0",
-        "@rollup/plugin-commonjs": "11.0.0",
-        "@rollup/plugin-node-resolve": "6.0.0",
-        "@semantic-release/changelog": "^5.0.1",
-        "@semantic-release/git": "^9.0.0",
-        "@sentry/types": "5.24.2",
-        "@types/bluebird": "3.5.29",
+        "@commitlint/cli": "^17.4.0",
+        "@commitlint/config-conventional": "^17.4.0",
+        "@rollup/plugin-babel": "^6.0.3",
+        "@rollup/plugin-commonjs": "^24.0.0",
+        "@rollup/plugin-node-resolve": "^15.0.1",
+        "@rollup/plugin-terser": "^0.3.0",
+        "@rollup/plugin-typescript": "^11.0.0",
+        "@semantic-release/changelog": "^6.0.2",
+        "@semantic-release/git": "^10.0.1",
         "@types/jest": "24.0.25",
-        "@types/js-base64": "2.3.1",
-        "@types/qs": "6.9.0",
-        "@typescript-eslint/eslint-plugin": "3.10.1",
-        "@typescript-eslint/parser": "3.10.1",
-        "@zerollup/ts-transform-paths": "1.7.9",
+        "@typescript-eslint/eslint-plugin": "^5.48.0",
+        "@typescript-eslint/parser": "^5.48.0",
+        "@zerollup/ts-transform-paths": "^1.7.18",
         "axios-mock-adapter": "1.17.0",
         "babel-eslint": "10.1.0",
         "babel-loader": "8.0.6",
         "babel-plugin-module-resolver": "4.0.0",
-        "commitizen": "^4.2.2",
-        "commitlint": "8.2.0",
-        "eslint": "7.8.1",
-        "eslint-config-prettier": "6.11.0",
-        "eslint-config-standard": "14.1.1",
-        "eslint-plugin-babel": "5.3.1",
-        "eslint-plugin-class-property": "^1.1.0",
-        "eslint-plugin-compat": "3.8.0",
-        "eslint-plugin-import": "2.22.0",
-        "eslint-plugin-json": "^2.1.2",
-        "eslint-plugin-node": "11.1.0",
-        "eslint-plugin-promise": "4.2.1",
-        "eslint-plugin-react": "7.20.6",
-        "eslint-plugin-react-hooks": "4.1.1",
-        "eslint-plugin-standard": "4.0.1",
-        "husky": "3.1.0",
-        "jest": "^26.6.3",
-        "jest-watch-typeahead": "^0.6.1",
+        "commitizen": "^4.2.6",
+        "cz-conventional-changelog": "^3.3.0",
+        "husky": "^8.0.3",
+        "jest": "^29.3.1",
+        "jest-watch-typeahead": "^2.2.1",
         "lint-staged": "9.5.0",
         "prettier": "2.1.1",
-        "qs": "6.9.1",
-        "remark-cli": "7.0.1",
-        "rollup": "1.27.14",
-        "rollup-plugin-babel": "4.3.3",
+        "remark-cli": "^11.0.0",
+        "rollup": "^3.9.1",
         "rollup-plugin-clear": "2.0.7",
-        "rollup-plugin-peer-deps-external": "2.2.0",
         "rollup-plugin-progress": "1.1.2",
-        "rollup-plugin-sourcemaps": "0.4.2",
-        "rollup-plugin-terser": "5.1.3",
-        "rollup-plugin-typescript2": "0.25.3",
-        "rollup-plugin-visualizer": "3.3.1",
-        "semantic-release": "^17.3.1",
-        "ts-jest": "^26.4.4",
-        "typedoc": "0.15.6",
-        "typescript": "3.7.4"
-    },
-    "peerDependencies": {
-        "qs": "6.9.1"
-    },
-    "config": {
-        "commitizen": {
-            "path": "cz-conventional-changelog"
-        }
+        "rollup-plugin-visualizer": "^5.9.0",
+        "semantic-release": "^20.0.0",
+        "ts-jest": "^29.0.3",
+        "ts-node": "^10.9.1",
+        "typedoc": "^0.23.24",
+        "typescript": "^4.9.4"
     }
 }
