+++ src/client/index.ts
@@ -1,3 +1,3 @@
-export * from './ApiClient'
+// export * from './ApiClient'
 export * from './pms'
 export * from './plex'
