+++ src/client/plex/index.ts
@@ -1,3 +1,3 @@
 export * from './Plex'
-export * from './users'
-export * from './resources'
+export * from './Users'
+export * from './Resources'
