--- src/client/plex/resources/index.ts
@@ -1,1 +1,0 @@
-export * from './Resources'
