--- src/client/pms/hubs/index.ts
@@ -1,1 +1,0 @@
-export * from './Hubs'
