+++ src/client/pms/index.ts
@@ -1,3 +1,3 @@
 export * from './PMS'
-export * from './library'
-export * from './hubs'
+export * from './Library'
+export * from './Hubs'
