--- src/client/pms/library/index.ts
@@ -1,1 +1,0 @@
-export * from './Library'
