+++ test/__fixtures__/library.getMetadata.200.json
