+++ test/__fixtures__/library.getMetadataMultiple.200.json
