+++ test/client/plex/Plex.spec.ts
@@ -1,4 +1,4 @@
-import { Plex } from '../Plex'
+import { Plex } from '@/client/plex'
 
 const fakeToken = 'fakeToken'
