+++ tsconfig.json
@@ -2,6 +2,8 @@
     "compilerOptions": {
         "sourceMap": true,
         "allowJs": false,
+        "declaration": true,
+        "outDir": "dist/lib",
         "module": "es2015",
         "moduleResolution": "node",
         "forceConsistentCasingInFileNames": true,
@@ -11,7 +13,6 @@
         "strictNullChecks": true,
         "suppressImplicitAnyIndexErrors": true,
         "noUnusedLocals": true,
-        "declaration": true,
         "esModuleInterop": true,
         "experimentalDecorators": true,
         "emitDecoratorMetadata": true,
@@ -20,16 +21,12 @@
         "rootDirs": ["src"],
         "baseUrl": ".",
         "paths": {
-            "@client": ["src/client"],
-            "@errors": ["src/errors"],
-            "@models": ["src/models"],
-            "@utils": ["src/utils"],
-            "@types": ["src/types"]
+            "@/*": ["src/*"],
+            "@fixtures/*": ["test/__fixtures__/*"]
         }
     },
-    "outDir": "build/lib",
     "target": "es5",
     "lib": ["es5", "es6", "es7", "esnext", "dom"],
-    "include": ["src/**/*", "**/__tests__", "tests/**/*"],
+    "include": ["src/**/*"],
     "exclude": ["node_modules"]
 }
